+++ src/dbd_helper.c
@@ -83,20 +83,20 @@ dbi_result_t *_dbd_result_create(dbi_conn_t *conn, void *handle, unsigned long l
 	return result;
 }
 
-void _dbd_result_set_numfields(dbi_result_t *result, unsigned short numfields) {
+void _dbd_result_set_numfields(dbi_result_t *result, unsigned long numfields) {
 	result->numfields = numfields;
 	result->field_names = calloc(numfields, sizeof(char *));
 	result->field_types = calloc(numfields, sizeof(unsigned short));
 	result->field_attribs = calloc(numfields, sizeof(unsigned int *));
 }
 
-void _dbd_result_add_field(dbi_result_t *result, unsigned short idx, char *name, unsigned short type, unsigned int attribs) {
+void _dbd_result_add_field(dbi_result_t *result, unsigned long idx, char *name, unsigned short type, unsigned int attribs) {
 	if (name) result->field_names[idx] = strdup(name);
 	result->field_types[idx] = type;
 	result->field_attribs[idx] = attribs;
 }
 
-dbi_row_t *_dbd_row_allocate(unsigned short numfields) {
+dbi_row_t *_dbd_row_allocate(unsigned long numfields) {
 	dbi_row_t *row = malloc(sizeof(dbi_row_t));
 	if (!row) return NULL;
 	row->field_values = calloc(numfields, sizeof(dbi_data_t));
@@ -152,7 +152,7 @@ void _dbd_internal_error_handler(dbi_conn_t *conn, const char *errmsg, const int
 dbi_result_t *_dbd_result_create_from_stringarray(dbi_conn_t *conn, unsigned long long numrows_matched, const char **stringarray) {
 	dbi_result_t *result = (dbi_result_t *) malloc(sizeof(dbi_result_t));
 	unsigned long long currow = 0;
-	const short numfields = 1;
+	const long numfields = 1;
 	
 	if (!result) return NULL;
